move resolver to a new file and fixed upon feedback
