+++ examples/utils/get-tinylicious-container/src/test/insecureTinyliciousUrlResolverTest.spec.ts
@@ -4,11 +4,9 @@
  */
 
 import { strict as assert } from "assert";
-// import { DriverHeader, IFluidResolvedUrl } from "@fluidframework/driver-definitions";
 import { ensureFluidResolvedUrl } from "@fluidframework/driver-utils";
-// import { IUser } from "@fluidframework/protocol-definitions";
 import { IRequest } from "@fluidframework/core-interfaces";
-import { InsecureTinyliciousUrlResolver } from "../getTinyliciousContainer";
+import { InsecureTinyliciousUrlResolver } from "../insecureTinyliciousUrlResolver";
 
 describe("Insecure Url Resolver Test", () => {
     const documentId = "fileName";
@@ -19,7 +17,7 @@ describe("Insecure Url Resolver Test", () => {
         resolver = new InsecureTinyliciousUrlResolver();
     });
 
-    it("Test RequestUrl for url with only document id", async () => {
+    it("Should resolve url with only document id", async () => {
         const testRequest: IRequest = {
             url: `${documentId}`,
             headers: {},
@@ -32,11 +30,10 @@ describe("Insecure Url Resolver Test", () => {
         assert.strictEqual(resolvedUrl.url, expectedResolvedUrl, "resolved url is wrong");
     });
 
-    it("Test RequestUrl for url with data object ids", async () => {
+    it("Should resolve url with data object ids", async () => {
         const path = "dataObject1/dataObject2";
         const testRequest: IRequest = {
             url: `${documentId}/${path}`,
-            headers: {},
         };
 
         const resolvedUrl = await resolver.resolve(testRequest);
@@ -46,10 +43,9 @@ describe("Insecure Url Resolver Test", () => {
         assert.strictEqual(resolvedUrl.url, expectedResolvedUrl, "resolved url is wrong");
     });
 
-    it("Test RequestUrl for url with a slash at the end", async () => {
+    it("Should resolve url with a slash at the end", async () => {
         const testRequest: IRequest = {
             url: `${documentId}/`,
-            headers: {},
         };
 
         const resolvedUrl = await resolver.resolve(testRequest);
@@ -59,10 +55,9 @@ describe("Insecure Url Resolver Test", () => {
         assert.strictEqual(resolvedUrl.url, expectedResolvedUrl, "resolved url is wrong");
     });
 
-    it("Test RequestUrl for url with 2 slashes at the end", async () => {
+    it("Should resolve url with 2 slashes at the end", async () => {
         const testRequest: IRequest = {
             url: `${documentId}//`,
-            headers: {},
         };
 
         const resolvedUrl = await resolver.resolve(testRequest);
@@ -72,17 +67,17 @@ describe("Insecure Url Resolver Test", () => {
         assert.strictEqual(resolvedUrl.url, expectedResolvedUrl, "resolved url is wrong");
     });
 
-    it("Test RequestUrl for url with special characters", async () => {
-        const path = "dataObject!@";
+    it("Should resolve url with special characters", async () => {
+        const path = "dataObject!@$";
+        const testDocumentId = "fileName!@$";
         const testRequest: IRequest = {
-            url: `${documentId}/${path}`,
-            headers: {},
+            url: `${testDocumentId}/${path}`,
         };
 
         const resolvedUrl = await resolver.resolve(testRequest);
         ensureFluidResolvedUrl(resolvedUrl);
 
-        const expectedResolvedUrl = `${hostUrl}/tinylicious/${documentId}/${path}`;
+        const expectedResolvedUrl = `${hostUrl}/tinylicious/${encodeURIComponent(testDocumentId)}/${path}`;
         assert.strictEqual(resolvedUrl.url, expectedResolvedUrl, "resolved url is wrong");
     });
 });
